+++ orc8r/cloud/go/tools/combine_swagger/generate/generate.go
@@ -11,7 +11,7 @@
  limitations under the License.
 */
 
-package gen
+package generate
 
 import (
 	"os"
