+++ orc8r/cloud/go/tools/combine_swagger/testdata/monolithic.yml.golden
